Automated g4 rollback of changelist 223158552.
*** Reason for rollback ***

Rollback of the Rollback in order to submit to github using copybara

*** Original change description ***

Automated g4 rollback of changelist 222891413.

*** Reason for rollback ***

Need to push to github with the CL unsubmitted

*** Original change description ***

Add an example with unicode character at the end of word.

cf https://groups.google.com/a/google.com/forum/#!topic/python-users/TjfwamTFuVI

Inauguration of the third_party to github syncs vs github to third_party sync.

***

***

PiperOrigin-RevId: 223162921